Refactor session list code and added a tabs
